+++ lib/combined_barcode_scanner.dart
@@ -1,6 +1,7 @@
 library combined_barcode_scanner;
 
+export 'src/barcode_scanner.dart';
 export 'src/scanner_configuration.dart';
 export 'src/scanner_properties.dart';
-export 'src/barcode_scanner.dart';
+export 'src/scanner_type.dart';
 export 'src/ui/barcode_scanner_widget.dart';